% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read.metadata}
\alias{read.metadata}
\title{tools Function}
\usage{
read.metadata(metadata.fname)
}
\arguments{
\item{metadata.fname}{Path to the metadata file}
}
\value{
noting
}
\description{
tools Function
}
\examples{
read.metadata()

}
